extract plist apple code into separate file
